junit5_app/src/main/java/org/juliangranda/junit5/ejemplo/models/Cuenta.java
package org.juliangranda.junit5.ejemplo.models;

import java.math.BigDecimal;
import java.util.Locale;

public class Cuenta {
    private String persona;
    private BigDecimal saldo;

    public Cuenta(String persona,BigDecimal saldo) {
        this.persona = persona.toUpperCase();
        this.saldo = saldo;
    }

    public String getPersona() {
        return persona;
    }

    public void setPersona(String persona) {
        this.persona = persona;
    }

    public BigDecimal getSaldo() {
        return saldo;
    }

    public void setSaldo(BigDecimal saldo) {
        this.saldo = saldo;
    }
}
